requirements.txt
agno
feedparser
pypdf
python-dotenv
streamlit
